// src/app/interfaces/i-scene.ts
export interface IScene {
  id: number;
  title: string;
  desc: string;
  decision: string;
  choice1: string;
  choice2: string;
  choice3: string;
  choice4: string;
  choice5: string;
  choice6: string;
  result1: number;
  result2: number;
  result3: number;
  result4: number;
  result5: number;
  result6: number;
  ending: boolean;
}